Changing /geo_data api to /data_update with refacto
