+++ tartare/tasks.py
@@ -1,16 +1,17 @@
 import glob
 import logging
 import shutil
+import os
+import datetime
 import zipfile
+from zipfile import ZipFile
+from shutil import copyfile
 
 from tartare import app
 from tartare import celery
-from shutil import copyfile
-import os
 from tartare.core import calendar_handler, models
 from tartare.core.calendar_handler import GridCalendarData
-from zipfile import ZipFile
-import datetime
+from tartare.core.data_handler import type_of_data, is_ntfs_data, is_calendar_data
 
 logger = logging.getLogger(__name__)
 
@@ -26,83 +27,6 @@ def update_data(coverage):
     input_dir = coverage.technical_conf.input_dir
     logger.info('scanning directory %s', input_dir)
     handle_data(input_dir, coverage.technical_conf.output_dir, coverage.technical_conf.current_data_dir)
-
-
-def type_of_data(filename):
-    """
-    return the type of data contains in a file + the path to load it
-    this type can be one in:
-     - 'gtfs'
-     - 'fusio'
-     - 'fare'
-     - 'osm'
-     - 'geopal'
-     - 'fusio'
-     - 'poi'
-     - 'synonym'
-     - 'shape'
-     if only_one_file is True, so consider only a zip for all pt data
-     else we consider also them for multi files
-    for 'fusio', 'gtfs', 'fares' and 'poi', we return the directory since there are several file to load
-    """
-    def files_type(files):
-        # first we try fusio, because it can load fares too
-        if any(f for f in files if f.endswith("contributors.txt")):
-            return 'fusio'
-        if any(f for f in files if f.endswith("grid_rel_calendar_to_network_and_line.txt")):
-            return 'calendar'
-        if any(f for f in files if f.endswith("fares.csv")):
-            return 'fare'
-        if any(f for f in files if f.endswith("stops.txt")):
-            return 'gtfs'
-        if any(f for f in files if f.endswith("adresse.txt")):
-            return 'geopal'
-        if any(f for f in files if f.endswith("poi.txt")):
-            return 'poi'
-        return None
-
-    if not isinstance(filename, list):
-        if os.path.isdir(filename):
-            files = glob.glob(filename + "/*")
-        else:
-            files = [filename]
-    else:
-        files = filename
-
-    # we test if we recognize a ptfile in the list of files
-    t = files_type(files)
-    if t:  # the path to load the data is the directory since there are several files
-        return t, os.path.dirname(files[0])
-
-    for filename in files:
-        if filename.endswith('.pbf'):
-            return 'osm', filename
-        if filename.endswith('.tmp'):
-            return 'tmp', filename
-        if filename.endswith('.zip'):
-            zipf = zipfile.ZipFile(filename)
-            pt_type = files_type(zipf.namelist())
-            if not pt_type:
-                return None, None
-            return pt_type, filename
-        if filename.endswith('.geopal'):
-            return 'geopal', filename
-        if filename.endswith('.poi'):
-            return 'poi', os.path.dirname(filename)
-        if filename.endswith("synonyms.txt"):
-            return 'synonym', filename
-        if filename.endswith(".poly") or filename.endswith(".wkt"):
-            return 'shape', filename
-
-    return None, None
-
-
-def is_ntfs_data(input_file):
-    return type_of_data(input_file)[0] == 'fusio'
-
-
-def is_calendar_data(input_file):
-    return type_of_data(input_file)[0] == 'calendar'
 
 
 def remove_old_ntfs_files(directory):
